Day7.
A couple of ugly hacks to speed it up or keep it short:
The program for day2 actually works for part1 (!).
Channels are not cleared - it is assumed the program consumes
everything.
The same codes program is not duplicated. Or in other words same copy
of the self modifying code runs on all machines. But it works (for me!).